Move gitcredentials HelperConfig and add tests
